Move config.js to the top level directory
