+++ src/main.js
@@ -2,7 +2,7 @@
 const mongoose = require('mongoose')
 const api = require('./api')
 const trends = require('./twitter/trends')
-const config = require('./config')
+const config = require('../config')
 const dbUtils = require('./utils/db-utils')
 const TweetStream = require('./twitter/tweet-stream')
 const tweetSearch = require('./twitter/tweet-search')
