+++ src/utils/sentiment-utils.js
@@ -1,5 +1,5 @@
 'use strict'
-var config = require('../config')
+var config = require('../../config')
 
 var sentimentUtils = {
